requirements.txt
numpy==1.15.2
pandas==0.24.2
pillow==5.4.1
tensorflow==1.13.1
